feat:Result 클래스에 구현
