--- src/main/java/Print/Percent.java
@@ -1,5 +1,0 @@
-package Print;
-
-public class Percent {
-
-}
