--- src/main/java/Print/Price.java
@@ -1,5 +1,0 @@
-package Print;
-
-public class Price {
-
-}
